drivers/cc2420: factorise/reorganise overall code
